Bootloader modificaiton (USB access still does not work)

+++ src/library/STM32_USB_HOST_Library/Class/MSC/src/usbh_msc_fatfs.c
@@ -183,7 +183,7 @@ DRESULT disk_ioctl (
     
   case GET_SECTOR_COUNT :	/* Get number of sectors on the disk (DWORD) */
     
-    *(DWORD*)buff = (DWORD) USBH_MSC_Param.MSCapacity;
+    *(uint32_t*)buff = USBH_MSC_Param.MSCapacity;
     res = RES_OK;
     break;
